refactor(compiler-sfc): extract defineEmits logic
